hw03/requirements.txt
streamlit
face_recognition
Pillow
numpy
pytest
